+++ NEWS.md
@@ -1,5 +1,6 @@
 # arcgisgeocode (development version)
 
+- The `world_geocoder` object has been deprecated in favor of `world_geocoder()` function
 - Resolves an issue parsing geocoding results with fields that are not present in the world geocoder. Fixes [#41](https://github.com/R-ArcGIS/arcgisgeocode/issues/41) <https://github.com/R-ArcGIS/arcgisgeocode/pull/42>
 
 # arcgisgeocode 0.3.0